+++ src/Validators/AbstractValidator.php
@@ -2,16 +2,70 @@
 
 namespace Rougin\Weasley\Validators;
 
-use Rougin\Weasley\Check;
+use Valitron\Validator;
 
 /**
- * @deprecated Use "Check" instead.
- *
  * Abstract Validator
  *
  * @package Weasley
  * @author  Rougin Gutib <rougingutib@gmail.com>
  */
-abstract class AbstractValidator extends Check
+abstract class AbstractValidator
 {
+    /**
+     * @var array<string, string[]>
+     */
+    public $errors = array();
+
+    /**
+     * @var \Valitron\Validator
+     */
+    protected $validator;
+
+    public function __construct()
+    {
+        $this->validator = new Validator;
+    }
+
+    /**
+     * Sets the labels in the validator.
+     *
+     * @return array<string, string>
+     */
+    abstract protected function labels();
+
+    /**
+     * Sets the rules in the validator.
+     *
+     * @param  array<string, mixed> $data
+     * @return void
+     */
+    abstract protected function rules(array $data = array());
+
+    /**
+     * Validates the given data against the specified rules.
+     *
+     * @param  array<string, mixed> $data
+     * @return boolean
+     */
+    public function validate(array $data)
+    {
+        $this->validator->labels($this->labels());
+
+        $this->rules($data);
+
+        $validator = $this->validator->withData($data);
+
+        if ($validator->validate())
+        {
+            return true;
+        }
+
+        /** @var array<string, string[]> */
+        $errors = $validator->errors();
+
+        $this->errors = $errors;
+
+        return false;
+    }
 }